CherryTree to Obsidian conversion
Conversion of the original notes from CherryTree into an Obsidian format. For the best results, place the four numbered markdown files into a folder named, "PenTesting Methodology" after downloading and then open that directory as a vault in Obsidian. That should allow each of the four files to be opened as a tab within the vault.

HTTP links were integrated using Obsidian's linking schema, but otherwise unaltered. Additional links were added throughout to repos for software when the software is first mentioned. Minor editing was performed to maintain readability after reformatting. Otherwise, original spellings and phrasing are preserved throughout.

Still being new to Obsidian, please feel free to make enhancements or corrections. I apologize in advance for any errors introduced in the process of converting the text. I have proofread the final version multiple times, but YMMV. There are a lot of good resources contained within for studying penetration testing.